ê오류 수정&경기상태조회

+++ src/main/java/com/GTTF/goal_to_the_future/domain/result/dto/response/RecordResultResponseDto.java
@@ -11,7 +11,7 @@ public class RecordResultResponseDto {
     private Integer yellowCard;
     private Integer redCard;
     private String highlight;
-    private Integer winnerTeamId;
+    private Long winnerTeamId;
     private Integer pass;
     private Integer effectiveShooting;
 }
